cgminer: Start work I/O and stage thread after all miner threads
This change prepares for allowing cgminer to run correctly even if
one or more miner threads can not be created.

Since the work I/O and stage thread data structures follow all miner
threads in the global array of threads, and since each thread is given
a pointer to their respective data structure, it is important not to
start any other threads until it is certain that the pointers will not
change.

A followup change gracefully handles the case that a miner thread can
not start, which requires all miner threads to be started first.