libztex: Rescan USB devices after firmware download
When one or more ZTEX devices with dummy firmware are found on startup, the
correct firmware is downloaded to each device, and then USB devices are
rescanned until all devices have reattached.

This allows to discover and configure all ZTEX devices and start mining
directly, without requiring a Java Runtime Environment and the ZTEX Java
software for sending the correct firmware to the device.